+++ include/fast_log/log_approx.h
@@ -7,16 +7,16 @@
 /// @returns the exponent and a normalized mantissa with the relationship:
 /// [a * 2^b] = x
 
-inline double logapprox(double val) {
+inline float logapprox(float val) {
   union {
-    double f;
+    float f;
     int32_t i;
   } valu;
-  double exp, addcst, x;
+  float exp, addcst, x;
   valu.f = val;
   exp = valu.i >> 23;
   /* -89.970756366f = -127 * log(2) + constant term of polynomial bellow. */
-  addcst = val > 0 ? -89.970756366f : -std::numeric_limits<double>::infinity();
+  addcst = val > 0 ? -89.970756366f : -std::numeric_limits<float>::infinity();
   valu.i = (valu.i & 0x7FFFFF) | 0x3F800000;
   x = valu.f;
 
@@ -27,9 +27,9 @@ inline double logapprox(double val) {
     > plot(f+(x-1)*log(2)-log(x), [1,2]);
     > f+(x-1)*log(2)
  */
-  return x * (3.529304993d +
+  return x * (3.529304993f +
               x * (-2.461222105f +
                    x * (1.130626167f +
-                        x * (-0.288739945f + x * 3.110401639e-2d)))) +
+                        x * (-0.288739945f + x * 3.110401639e-2f)))) +
          (addcst + 0.6931471805f * exp);
 }
