+++ include/panel_geo/panel_geo.hpp
@@ -10,8 +10,8 @@ template <SurfaceType T> struct PanelGeometry;
 
 namespace PanelGeometryUtils {
 
-Eigen::MatrixXd rowwiseCross(const Eigen::MatrixXd &A,
-                             const Eigen::MatrixXd &B);
+Eigen::MatrixXf rowwiseCross(const Eigen::MatrixXf &A,
+                             const Eigen::MatrixXf &B);
 
 } // namespace PanelGeometryUtils
 
@@ -21,7 +21,7 @@ public:
   using Scalar = typename SurfacePanel::NodeMatrix::Scalar;
   using PointType = Eigen::Array<Scalar, 1, 3>;
   using VecType = PointType;
-  // Constrain with concepts to be an Eigen indexable type: arrayXd, array,
+  // Constrain with concepts to be an Eigen indexable type: arrayXf, array,
   // vector
 
   SurfacePanel::NodeMatrix centrePoints;
@@ -29,9 +29,9 @@ public:
   SurfacePanel::NodeMatrix tangentYVectors;
   SurfacePanel::NodeMatrix normalVectors;
 
-  Eigen::ArrayXd areas;
-  std::vector<Eigen::Isometry3d> conversionMatrices;
-  std::vector<Eigen::ArrayX3d> localFaceVertices;
+  Eigen::ArrayXf areas;
+  std::vector<Eigen::Isometry3f> conversionMatrices;
+  std::vector<Eigen::ArrayX3f> localFaceVertices;
 
   T mSurface;
 
@@ -41,9 +41,9 @@ public:
   PanelGeometry(T &&surface) noexcept;
   void panelGeoInit();
   void calculateCentrePointsandVectors();
-  Eigen::Isometry3d createLocalConversionMatrix(std::size_t faceIdx);
-  Eigen::ArrayX3d convertToLocal(int faceIdx, const ArrayX3d &points) const;
-  double calcPolyArea(const Eigen::ArrayX3d &vertices) const;
+  Eigen::Isometry3f createLocalConversionMatrix(std::size_t faceIdx);
+  Eigen::ArrayX3f convertToLocal(int faceIdx, const ArrayX3f &points) const;
+  float calcPolyArea(const Eigen::ArrayX3f &vertices) const;
 };
 
 using PanelGeometryPair =
