+++ include/solver/dense_gmres_solver.hpp
@@ -11,7 +11,7 @@
 #include <unsupported/Eigen/SparseExtra>
 #include <vector>
 // #define ANALYSIS_DIR
-// "D:/PortableDev/projects/panel_methods_3d/python/out_cpp"
+// "D:/PortableDev/projects/panel_methods_3f/python/out_cpp"
 
 #define SAVE_SYSTEM 0
 #define ITER_RES 0
@@ -19,22 +19,22 @@ using namespace Eigen;
 
 struct DenseGMRESSolver : ISolver<DenseGMRESSolver> {
 
-  inline static double spTol = 1e-6;
+  inline static float spTol = 1e-6;
 
 public:
-  auto setspTol(double pTol_) {
+  auto setspTol(float pTol_) {
     spTol = pTol_;
     return *this;
   }
   DenseGMRESSolver() : ISolver<DenseGMRESSolver>() {}
   template <typename MatrixType, typename VecType>
-  static Eigen::VectorXd solveImpl(const Eigen::MatrixBase<MatrixType> &lhs,
+  static Eigen::VectorXf solveImpl(const Eigen::MatrixBase<MatrixType> &lhs,
                                    const Eigen::MatrixBase<VecType> &rhs,
-                                   double tol = 1e-6,
+                                   float tol = 1e-6,
                                    std::size_t maxit = 1000) {
     // std::cout << "Creating...\n";
-    using SpMat = Eigen::SparseMatrix<double>;
-    Eigen::SparseMatrix<double> precond_mat(lhs.rows(), lhs.cols());
+    using SpMat = Eigen::SparseMatrix<float>;
+    Eigen::SparseMatrix<float> precond_mat(lhs.rows(), lhs.cols());
     precond_mat = lhs.sparseView(1e-2, 1);
 
     const Eigen::IdentityPreconditioner preconditioner;
@@ -42,7 +42,7 @@ public:
 
     // printf("DropTol: %1.6f\n", dropTol);
     // std::cout << "Solving...\n";
-    Eigen::VectorXd x(rhs.rows());
+    Eigen::VectorXf x(rhs.rows());
     x.setZero();
     Eigen::Index iters = 1000;
     Eigen::internal::gmres(lhs, rhs, x, preconditioner, iters, maxit, tol);
