+++ include/solver/gmres_solver.hpp
@@ -11,28 +11,28 @@
 #include <unsupported/Eigen/SparseExtra>
 #include <vector>
 // #define ANALYSIS_DIR
-// "D:/PortableDev/projects/panel_methods_3d/python/out_cpp"
+// "D:/PortableDev/projects/panel_methods_3f/python/out_cpp"
 
 #define SAVE_SYSTEM 0
 #define ITER_RES 0
 
 struct GMRESSolver : ISolver<GMRESSolver> {
-  inline static double spTol = 1e-6;
+  inline static float spTol = 1e-6;
 
 public:
   GMRESSolver() : ISolver<GMRESSolver>() {}
-  auto setspTol(double spTol_) {
+  auto setspTol(float spTol_) {
     spTol = spTol_;
     return *this;
   }
   template <typename MatrixType, typename VecType>
-  static Eigen::VectorXd solveImpl(const Eigen::MatrixBase<MatrixType> &lhs,
+  static Eigen::VectorXf solveImpl(const Eigen::MatrixBase<MatrixType> &lhs,
                                    const Eigen::MatrixBase<VecType> &rhs,
-                                   double tol = 1e-6,
+                                   float tol = 1e-6,
                                    std::size_t maxit = 1000) {
     // std::cout << "Creating...\n";
-    typedef Eigen::SparseMatrix<double> SpMat;
-    typedef Eigen::Triplet<double> T;
+    typedef Eigen::SparseMatrix<float> SpMat;
+    typedef Eigen::Triplet<float> T;
 
     SpMat A(lhs.rows(), lhs.cols());
     A = lhs.sparseView(spTol, 1);
@@ -43,7 +43,7 @@ public:
     // printf("\n");
     // printf("%1.10f\n", spTol);
     // printf("%1.6f\n",
-    //        ((double)A.nonZeros()) / ((double)(lhs.rows() * lhs.cols())));
+    //        ((float)A.nonZeros()) / ((float)(lhs.rows() * lhs.cols())));
 
     // printf("DropTol: %1.6f\n", dropTol);
     const Eigen::IdentityPreconditioner preconditioner;
@@ -51,7 +51,7 @@ public:
 
     // printf("DropTol: %1.6f\n", dropTol);
     // std::cout << "Solving...\n";
-    Eigen::VectorXd x(rhs.rows());
+    Eigen::VectorXf x(rhs.rows());
     x.setZero();
     Eigen::Index iters = 1000;
     Eigen::internal::gmres(lhs, rhs, x, preconditioner, iters, maxit, tol);
